requirements.txt
Flask
numpy
opencv-python
Pillow
scikit-image
rembg